% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dbfishR.R
\name{fetch_dbfishR}
\alias{fetch_dbfishR}
\title{Retrieve FRI data}
\usage{
fetch_dbfishR(
  url = getOption("dbfishR_data_url"),
  cache = getOption("dbfishR_use_cache"),
  cache_dir = getOption("dbfishR_cache_dir"),
  refresh = getOption("dbfishR_refresh"),
  show_progress = getOption("dbfishR_progress")
)
}
\arguments{
\item{url}{URL for FRI's AccessDB database.}

\item{cache}{Whether to cache the AccessDB for future R sessions.}

\item{cache_dir}{The location to cache the AccessDB.}

\item{refresh}{Whether to force a re-download the AccessDB.}

\item{show_progress}{If set to FALSE, do not display download progress bar (helpful for R Markdown documents).}
}
\value{
The local path to FRI's AccessDB
}
\description{
Helper function to download FRI's data
}
\keyword{internal}
