% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dbfishR.R
\name{fetch_dbfishR_template}
\alias{fetch_dbfishR_template}
\title{Retrieve FRI's data template}
\usage{
fetch_dbfishR_template(
  url = getOption("dbfishR_template_url"),
  cache = getOption("dbfishR_use_cache"),
  cache_dir = getOption("dbfishR_cache_dir"),
  refresh = getOption("dbfishR_refresh"),
  show_progress = getOption("dbfishR_progress")
)
}
\arguments{
\item{url}{URL for FRI's data template.}

\item{cache}{Whether to cache the template for future R sessions.}

\item{cache_dir}{The location to cache the template.}

\item{refresh}{Whether to force a re-download the template.}

\item{show_progress}{If set to FALSE, do not display download progress bar (helpful for R Markdown documents).}
}
\value{
The local path to FRI's Template
}
\description{
Helper function to download FRI's data template so you can contribute
}
\keyword{internal}
