% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_discharge_event_info}
\alias{get_discharge_event_info}
\title{Retrieves discharge events}
\usage{
get_discharge_event_info(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{\link[dbfishR]{fetch_dbfishR}}}
}
\value{
\code{\link[base]{data.frame}}
}
\description{
Retrieve discharge event data from the FRI database
}
\examples{

\dontrun{
library(dbfishR)
data <- get_discharge_event_info()
View(data)
}

}
