% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_events}
\alias{get_events}
\title{get_events}
\usage{
get_events(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{\link[dbfishR]{fetch_dbfishR}}}
}
\description{
retrieves the events.
}
\examples{

\dontrun{
library(dbfishR)
data <- get_events()
View(data)
}

}
