% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_fish_survey}
\alias{get_fish_survey}
\title{Retrieve the fish survey}
\usage{
get_fish_survey(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{\link[dbfishR]{fetch_dbfishR}}}
}
\value{
\code{\link[base]{data.frame}}
}
\description{
Retrieve fish survey data from the \href{https://www.susqu.edu/academics/centers-and-programs/freshwater-research-institute}{FRI} database
}
\examples{

\dontrun{
library(dbfishR)
data <- get_fish_survey()
View(data)
}

}
