% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_particle_size}
\alias{get_particle_size}
\title{get_particle_size}
\usage{
get_particle_size(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{\link[dbfishR]{fetch_dbfishR}}}
}
\description{
retrieves the particle size.
}
\examples{

\dontrun{
library(dbfishR)
data <- get_particle_size()
View(data)
}

}
