% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_sites}
\alias{get_sites}
\title{Retrieve the sites}
\usage{
get_sites(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{\link[dbfishR]{fetch_dbfishR}}}
}
\value{
\code{\link[base]{data.frame}}
}
\description{
Retrieve site data from the \href{https://www.susqu.edu/academics/centers-and-programs/freshwater-research-institute}{FRI} database
}
\examples{

\dontrun{
library(dbfishR)
data <- get_sites()
View(data)
}

}
