% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{get_table}
\usage{
get_table(table, ...)
}
\arguments{
\item{table}{the table to fetch from the db}

\item{...}{additional parameters to pass to \code{\link[dbfishR]{fetch_dbfishR}}}
}
\value{
xxx
}
\description{
xxx
}
