% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_events}
\alias{get_events}
\title{Retrieve the events}
\usage{
get_events(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{\link[dbfishR]{fetch_dbfishR}}}
}
\value{
\code{\link[base]{data.frame}}
}
\description{
Retrieve event data from the \href{https://www.susqu.edu/academics/centers-and-programs/freshwater-research-institute}{FRI} database
}
\examples{

\dontrun{
library(dbfishR)
data <- get_events()
View(data)
}

}
