% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_fish_records}
\alias{get_fish_records}
\title{get_fish_records}
\usage{
get_fish_records(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{\link[dbfishR]{fetch_dbfishR}}}
}
\description{
retrieves the fish records.
}
\examples{

\dontrun{
library(dbfishR)
data <- get_fish_records()
View(data)
}

}
