% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_fish_survey}
\alias{get_fish_survey}
\title{get_fish_survey}
\usage{
get_fish_survey(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{\link[dbfishR]{fetch_dbfishR}}}
}
\description{
retrieves the fish survey.
}
