% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_fish_tally}
\alias{get_fish_tally}
\title{get_fish_tally}
\usage{
get_fish_tally(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{\link[dbfishR]{fetch_dbfishR}}}
}
\description{
retrieves the fish tally.
}
