% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_sites}
\alias{get_sites}
\title{get_sites}
\usage{
get_sites(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{\link[dbfishR]{fetch_dbfishR}}}
}
\description{
retrieves the sites.
}
